Update human clinical and health data sources page (#93)
* Update CONTRIBUTORS.yaml

* Update human clinical and health data sources page

As we were encouraged to get familiar with GitHub I'm trying to add a first few lines of text to the human clinical and health data sources page. This is a first draft to try everything out, but hopefully it's a start.

* Add image for human clinical and health data sources page (#92)

As we were encouraged to get familiar with GitHub I'm trying to add a first few lines of text and an image to the human clinical and health data sources page.

* Update data_sources.md

* Update data_sources.md

Aligned page with latest text Google Drive

* Update data_sources.md

The changes suggested by Shona have been made and the page reorganised. Currently I'm not yet happy with the section 'Finding (meta)data - Considerations'. Any suggestions?

* Update data_sources.md

This update addresses the latest feedback received from Shona and Laura. If all is well the page should be near ready to be published.

* add file

* fix list affiliations

* no afifliations on non-showcase pages

* first changes to data sources health page

* changes to health data sources page

* Small change to the health data sources page

* Extra small changes to the health data sources page

* Add image to the health data sources page

* shorten Beacon description

* adding me as contributor

* add related pages

* add news item

* change style of the lists

---------

Co-authored-by: Stian Soiland-Reyes <soiland-reyes@manchester.ac.uk>
Co-authored-by: bedroesb <bedro@psb.vib-ugent.be>
Co-authored-by: Laura Portell Silva <74184187+lauportell@users.noreply.github.com>